\encoding{latin1}
\name{summary.glmmboot}
\alias{summary.glmmboot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Summary of a glmmboot object}
\description{
It simply calls \code{print.glmmboot}
}
\usage{
\method{summary}{glmmboot}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A glmmboot object}
  \item{\dots}{Additional arguments}
}
\details{
  A summary method will be written soon.
}
\value{
Nothing is returned.
}
%\references{See }
\author{Göran Broström}
\note{Preliminary}

 
\seealso{\code{\link{print.glmmboot}}}
%\examples{
%
%}
\keyword{print}% at least one, from doc/KEYWORDS

