\name{summary.glmmML}
\alias{summary.glmmML}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Summary of a glmmML object}
\description{
It simply calls \code{print.glmmML}
}
\usage{
summary.glmmML(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A glmmML object}
  \item{\dots}{Additional arguments}
}
\details{
  A summary method will be written soon.
}
\value{
Nothing is returned.
}
\references{See }
\author{Göran Broström}
\note{Preliminary}

 
\seealso{\code{\link{print.glmmML}}}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(object, ...){
    print.glmmML(object, ...)
  }
}
\keyword{print}% at least one, from doc/KEYWORDS

