\encoding{UTF-8}
\name{print.glmmML}
\alias{print.glmmML}
\title{Prints a 'glmmML' object.}
\description{
A glmmML object is the output of \code{glmmML}.
}
\usage{
\method{print}{glmmML}(x, digits = max(3, getOption("digits") - 3), na.print = "", ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{x}{The glmmML object}
  \item{digits}{Number of printed digits.}
  \item{na.print}{How to print NAs}
  \item{\dots}{Additional parameters, which are ignored.}
}
\details{
Nothing in particular.
}
\value{
  A short summary of the object is printed.
}
%\references{\enc{Broström}{Brostrom} (2003)}
\author{\enc{Göran Broström}{Goran Brostrom}}
\note{This is the only summary method available for the moment.}

\seealso{\code{\link{glmmML}}}

%\examples{
%
%}
\keyword{print}% at least one, from doc/KEYWORDS
