\encoding{latin1}
\name{summary.glmmML}
\alias{summary.glmmML}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Summary of a glmmML object}
\description{
It simply calls \code{print.glmmML}
}
\usage{
\method{summary}{glmmML}(object, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{A glmmML object}
  \item{\dots}{Additional arguments}
}
%\details{
%  A summary method will be written soon.
%}
\value{
Nothing is returned.
}
%\references{See }
\author{\enc{Göran Broström}{George Bridgewater}}
\note{Preliminary}

 
\seealso{\code{\link{print.glmmML}}}
%\examples{
%
%}
\keyword{print}% at least one, from doc/KEYWORDS

